+++ src/torchmetrics/functional/clustering/__init__.py
@@ -11,6 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from torchmetrics.functional.clustering.adjusted_rand_score import adjusted_rand_score
 from torchmetrics.functional.clustering.calinski_harabasz_score import calinski_harabasz_score
 from torchmetrics.functional.clustering.dunn_index import dunn_index
 from torchmetrics.functional.clustering.mutual_info_score import mutual_info_score
@@ -18,6 +19,7 @@ from torchmetrics.functional.clustering.normalized_mutual_info_score import norm
 from torchmetrics.functional.clustering.rand_score import rand_score
 
 __all__ = [
+    "adjusted_rand_score",
     "calinski_harabasz_score",
     "dunn_index",
     "mutual_info_score",